requirements.txt
fastapi
uvicorn[standard]
gunicorn
SQLAlchemy>=2.0.0
psycopg2-binary
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
httpx
jinja2